refactor: 오브제 도메인 패키지 구조 레이어드 아키텍처로 변경
