+++ src/main/java/com/example/daobe/objet/domain/ObjetStatus.java
@@ -1,4 +1,4 @@
-package com.example.daobe.objet.entity;
+package com.example.daobe.objet.domain;
 
 public enum ObjetStatus {
     ACTIVE,     // 사용 중인 상태
