+++ src/main/java/com/example/daobe/objet/domain/repository/ObjetRepository.java
@@ -1,7 +1,7 @@
-package com.example.daobe.objet.repository;
+package com.example.daobe.objet.domain.repository;
 
-import com.example.daobe.objet.entity.Objet;
-import com.example.daobe.objet.entity.ObjetStatus;
+import com.example.daobe.objet.domain.Objet;
+import com.example.daobe.objet.domain.ObjetStatus;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
